Rename to sharing to allow for other secret sharing methods
