+++ libursa/Cargo.toml
@@ -62,7 +62,7 @@ path = "bin/test_xchacha20poly1305.rs"
 required-features = ["benchmarkxchacha20poly1305"]
 
 [features]
-default = ["encryption", "ffi", "hashes", "kex", "serde", "signatures_native", "shamir_native"]
+default = ["encryption", "ffi", "hashes", "kex", "serde", "signatures_native", "sharing_native"]
 aescbc = ["aead", "aes", "block-modes", "block-padding", "hex", "hmac", "rand", "sha2/std", "subtle", "zeroize"]
 aescbc_native = ["aead", "hex", "rand", "openssl", "zeroize"]
 aesgcm_native = ["aead", "hex", "rand", "openssl", "zeroize"]
@@ -97,10 +97,10 @@ kex = ["ecdh_secp256k1", "x25519"]
 kex_native = ["ecdh_secp256k1_native", "x25519"]
 kex_asm = ["ecdh_secp256k1_asm", "x25519_asm"]
 logger = ["env_logger", "log"]
-portable = ["clear_on_drop/no_cc", "encryption", "hashes", "kex", "serde", "signatures"]
+portable = ["clear_on_drop/no_cc", "encryption", "hashes", "kex", "serde", "signatures", "sharing"]
 portable_wasm = ["portable", "wasm"]
-shamir = ["failure", "glass_pumpkin", "int_traits", "lazy_static", "num-bigint", "num-integer", "num-traits", "log", "rand", "sha2/std", "time"]
-shamir_native = ["failure", "int_traits", "lazy_static", "log", "openssl", "rand", "time"]
+sharing = ["failure", "glass_pumpkin", "int_traits", "lazy_static", "num-bigint", "num-integer", "num-traits", "log", "rand", "sha2/std", "time"]
+sharing_native = ["failure", "int_traits", "lazy_static", "log", "openssl", "rand", "time"]
 signatures = ["cl", "ed25519", "ecdsa_secp256k1", "bls_bls12381", "bls_bn254"]
 signatures_native = ["cl_native", "ed25519", "ecdsa_secp256k1_native", "bls_bls12381", "bls_bn254"]
 signatures_asm = ["cl_native", "ed25519_asm", "ecdsa_secp256k1_asm", "bls_bls12381", "bls_bn254_asm"]
