+++ libursa/src/sharing/mod.rs
